refactoring code
Im applying what i learned from the infallable code youtube channel to clean up my code, starting with the player. theres minimal risk with this thanks to github and i should be able to

A.) improve performance and
B.) improve readiability

I also switched over to jetbrains rider to test it out

So For Physics controller I updated all of the functions, (and rolled some together) and re-named a lot of stuff, I also made a lot more stuff private so it cant be changed from outside of the class.

I also made some changes to stepToggle.cs but that seems to be broken at the moment and needs more exploration, honestly i might go back to visual studio, im just more comfortable there...

But with that i think im done for the day, I want to take a second look at refactoring because i think there are still more readability and speed gains to get from it but we will have to see.